+++ docs/Changelog.md
@@ -9,6 +9,7 @@
 - [bugfix] active timers, if any, are cancelled upon actor shutdown finish
 - [bugfix] supervisor shutdown message is lost in rare cases right after 
 child actor start
+- [example] `examples/thread/sha512.cpp` (new)
 - [documentation] updated `Event loops & platforms`
 - [deprecated] state_response_t, state_request_t will be removed in v0.13